+++ netket/hilbert/index/constraints/__init__.py
@@ -1,0 +1,2 @@
+from .base import optimalConstrainedHilbertindex, ConstrainedHilbertIndex
+from .sum import SumConstraint