+++ netket/hilbert/qubit.py
@@ -14,6 +14,7 @@
 
 from typing import Optional, Union
 
+import numpy as np
 
 from netket.utils import StaticRange
 
@@ -37,7 +38,7 @@ class Qubit(HomogeneousHilbert):
             >>> print(hi.size)
             100
         """
-        super().__init__(StaticRange(0, 1, 2, dtype=float), N)
+        super().__init__(StaticRange(0, 1, 2, dtype=np.int8), N)
 
     def __pow__(self, n):
         return Qubit(self.size * n)